requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
beautifulsoup4==4.12.2
billiard==4.2.1
cachetools==5.5.2
celery==5.3.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==4.2.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.3
django-timezone-field==7.1
djangorestframework==3.14.0
feedparser==6.0.10
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.62.3
httplib2==0.31.0
idna==3.10
kombu==5.5.4
lxml==4.9.3
numpy==2.3.3
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-crontab==3.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
rsa==4.9.1
setuptools==78.1.1
sgmllib3k==1.0.0
shapely==2.1.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
